// frontend/src/client/models/CategoryResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Category response schema
 */
export type CategoryResponse = {
    name: string;
    slug: string;
    description?: (string | null);
    image_url?: (string | null);
    parent_id?: (number | null);
    id: number;
};